fix: api tokens hook
