+++ packages/web/src/hooks/stellar/tokens/index.ts
@@ -1,8 +1,6 @@
 export * from './utils';
-export * from './use-lp-tokens';
 export * from './use-trustline';
 export * from './use-api-tokens';
-export * from './use-all-tokens';
 export * from './use-token-price';
 // export * from './use-user-tokens';
 export * from './use-token-balance';
