+++ packages/web/src/hooks/stellar/tokens/use-api-tokens.ts
@@ -1,26 +1,39 @@
-import type { StateToken, TokenMapType } from '@normalfinance/types';
+import type { ApiToken } from '@normalfinance/types';
 
+import axios from 'axios';
 import { useState, useEffect } from 'react';
 import useSWRImmutable from 'swr/immutable';
-
-import { tokensToMap } from './utils';
-import { fetchApiTokens } from './use-featured-tokens';
+import { constants } from '@normalfinance/utils';
 
 //Returns tokens from the API
 export const useApiTokens = () => {
-  const { data, mutate, isLoading, error } = useSWRImmutable(['tokens'], () => fetchApiTokens());
+  const { data, mutate, isLoading, error } = useSWRImmutable(['api_tokens'], () =>
+    fetchApiTokens()
+  );
 
-  const [tokens, setTokens] = useState<StateToken[]>([]);
-  const [tokensAsMap, setTokensAsMap] = useState<TokenMapType>({});
+  const [tokens, setTokens] = useState<ApiToken[]>([]);
+  // const [tokensAsMap, setTokensAsMap] = useState<TokenMapType>({});
 
   useEffect(() => {
-    if (data && data.length > 0) {
+    if (data) {
       setTokens(data.assets);
     }
 
-    const mappedTokens = tokensToMap(tokens);
-    setTokensAsMap(mappedTokens);
+    // const mappedTokens = tokensToMap(tokens);
+    // setTokensAsMap(mappedTokens);
   }, [data, tokens]);
 
-  return { tokens, mutate, isLoading, isError: error, data, tokensAsMap };
+  return { tokens, mutate, isLoading, isError: error, data };
+};
+
+export const fetchApiTokens = async () => {
+  let url = 'https://raw.githubusercontent.com/normalfinance/token-list/main/tokenList.json';
+
+  if (constants.RPC_URL.includes('testnet')) {
+    url = 'https://raw.githubusercontent.com/normalfinance/token-list/main/tokenListTestnet.json';
+  }
+
+  const { data } = await axios.get(url);
+
+  return data;
 };
